requirements.txt
asgiref==3.5.0 
Django==4.0.3 
djongo==1.3.6 
dnspython==2.2.1 
pymongo==3.12.1 
python-snappy==0.6.1 
pytz==2022.1 
sqlparse==0.2.4
django-cron-0.6.0
